Fixed wrong conflict resolution - should have left he main branch code

+++ cm/util/misc.py
@@ -568,12 +568,8 @@ def get_file_from_bucket(conn, bucket_name, remote_filename, local_file, validat
         except S3ResponseError as e:
             log.debug("Failed to get file '%s' from bucket '%s': %s" % (
                 remote_filename, bucket_name, e))
-            try:
-                os.remove(local_file) # Don't leave a partially downloaded or touched file
-            except:
-                pass # Just in case the file does not exist at all (we still want to return False not an exception)   
-            return False
-    else:
+            if os.path.exists(local_file):
+                os.remove(local_file)  # Don't leave a partially downloaded or touched file    else:
         log.debug("Bucket '%s' does not exist, did not get remote file '%s'" % (
             bucket_name, remote_filename))
         return False
